==> analysis/01_process_expression.Rmd <==
---
title: "Process expression"
site: workflowr::wflow_site
author: "Julien Bryois"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  workflowr::wflow_html:
   toc: true
   toc_float: true
editor_options:
  chunk_output_type: console
---

```{r,echo=FALSE}
#Necessary for module load to work
knitr::opts_chunk$set(engine.opts = list(bash="--login"))
```

```{r,message=FALSE}
library(tidyverse)
library(SingleCellExperiment)
library(parallel)
```

# Files

Get individuals with both scRNA-seq and genotype information

```{r,eval=TRUE,message=FALSE}
individidual_genotype_scRNA <- read_tsv('data_sensitive/genotypes/individuals_with_genotype_scRNA.txt',
                                        col_names = FALSE)
```

# MS expression

```{r,eval=TRUE}
sce_ms <- readRDS('data_sensitive/sce/ms_sce_3.rds')
```

get annotation

```{r,eval=TRUE,message=FALSE}
annot <- read_csv('data_sensitive/sce/conos_labelled_2021-05-31.txt')
```

Only keep cells that were annotated

```{r,eval=TRUE}
sce_ms <- sce_ms[,colnames(sce_ms)%in%annot$cell_id]
```

```{r,eval=TRUE}
m <- match(colnames(sce_ms),annot$cell_id)
sce_ms$broad_annot <- annot$type_broad[m]
```

Some scRNA-seq were swapped and not corrected in the loaded sce object.

EU017 and EU018 --> swap
WM189 and WM190 --> swap

Swap them here

```{r,eval=TRUE}
sce_ms$library_id[sce_ms$library_id=='EU017'] <- 'tmp'
sce_ms$library_id[sce_ms$library_id=='EU018'] <- 'EU017'
sce_ms$library_id[sce_ms$library_id=='tmp'] <- 'EU018'

colnames(sce_ms) <- gsub('EU017','tmp',colnames(sce_ms))
colnames(sce_ms) <- gsub('EU018','EU017',colnames(sce_ms))
colnames(sce_ms) <- gsub('tmp','EU018',colnames(sce_ms))
```

```{r,eval=TRUE}
sce_ms$library_id[sce_ms$library_id=='WM189'] <- 'tmp'
sce_ms$library_id[sce_ms$library_id=='WM190'] <- 'WM189'
sce_ms$library_id[sce_ms$library_id=='tmp'] <- 'WM190'

colnames(sce_ms) <- gsub('WM189','tmp',colnames(sce_ms))
colnames(sce_ms) <- gsub('WM190','WM189',colnames(sce_ms))
colnames(sce_ms) <- gsub('tmp','WM190',colnames(sce_ms))
```

Keep only individuals for which we have both scRNA-seq and expression

```{r,eval=TRUE}
sce_ms <- sce_ms[,sce_ms$patient_id%in%individidual_genotype_scRNA$X1]
```

add individual_id column to the sce object

```{r,eval=TRUE}
sce_ms$individual_id <- sce_ms$patient_id
```

Let's remove the individual which is also in the AD dataset (less cells than the AD dataset)

```{r,eval=TRUE}
sce_ms <- sce_ms[,sce_ms$individual_id!='S03-009']
```

Let's remove samples that don't match their genotype

```{r,eval=TRUE}
sce_ms <- sce_ms[,!sce_ms$library_id%in%c('EU015','EU004','WM115','WM155','WM185')]
```

Let's remove immune cells from the MS dataset (not observed in the AD dataset)

```{r,eval=TRUE}
sce_ms <- sce_ms[,!sce_ms$broad_annot=='Immune']
```

Let's set the gene name to ensembl (to merge with the AD sce dataset)

```{r,eval=TRUE}
rownames(sce_ms) <- rowData(sce_ms)$ensembl
```

Finally, let's save info about which sample corresponds to which individual in the filtered data

```{r,eval=TRUE}
ms_sample_individual <- colData(sce_ms) %>% 
  as_tibble() %>% 
  dplyr::select(sample_id=library_id,individual_id,age,post_mortem_m,sex,tissue=matter) %>% 
  unique() %>% 
  write_tsv(.,'data_sensitive/expression/sample_individual_ms.txt',col_names = TRUE)
```

# AD expression

```{r,eval=TRUE}
sce_ad <- readRDS('data_sensitive/sce/sce.annotated8.rds')
```

Keep only individuals for which we have both scRNA-seq and expression

```{r,eval=TRUE}
sce_ad <- sce_ad[,sce_ad$individual_id%in%individidual_genotype_scRNA$X1]
```

Some individuals are present in multiple AD studies

```{r,eval=TRUE}
(individuals_multiple_studies <- colData(sce_ad) %>% as_tibble() %>% 
  dplyr::select(individual_id,study,diagnosis) %>% 
  unique() %>% 
  add_count(individual_id) %>% 
  filter(n>1) %>% 
  arrange(individual_id))
```

count number of cells for these individuals

```{r,eval=TRUE}
n_cells_per_id <- colData(sce_ad) %>% as_tibble() %>% 
  filter(individual_id%in%individuals_multiple_studies$individual_id) %>% 
  dplyr::count(individual_id,study) %>% 
    group_by(individual_id) %>% 
    mutate(keep=ifelse(n==max(n),TRUE,FALSE))
```

```{r,eval=TRUE}
to_remove <- filter(n_cells_per_id,!keep) %>% 
  mutate(lab=paste0(individual_id,'_',study))
```

```{r,eval=TRUE}
sce_ad$lab <- paste0(sce_ad$individual_id,'_',sce_ad$study)
```

Remove the sample with the least amount of cells (when two samples from the same individual are observed in two datasets)

```{r,eval=TRUE}
sce_ad <- sce_ad[,!sce_ad$lab%in%to_remove$lab]
```

Get same cell type label as the MS dataset

```{r,eval=TRUE}
sce_ad$broad_annot <- case_when(
  sce_ad$broad_annot=='Glutamatergic' ~ 'Excitatory neurons',
  sce_ad$broad_annot=='GABAergic' ~ 'Inhibitory neurons',
  sce_ad$broad_annot=='OPCs' ~ 'OPCs / COPs',
  sce_ad$broad_annot=='Endothelial' ~ 'Endothelial cells',
  TRUE ~ sce_ad$broad_annot)
```

Let's set the gene name to ensembl (to merge with the AD sce dataset)

```{r,eval=TRUE}
rownames(sce_ad) <- rowData(sce_ad)$ENSEMBL
```

Finally, let's save info about which sample corresponds to which individual in the filtered data

```{r,eval=TRUE}
ad_sample_individual <- colData(sce_ad) %>% 
  as_tibble() %>% 
  mutate(post_mortem_m=pmi*60) %>% 
  dplyr::select(sample_id,individual_id,age,post_mortem_m,sex,tissue) %>% 
  unique() %>% 
  write_tsv(.,'data_sensitive/expression/sample_individual_ad.txt',col_names = TRUE)
```

# Get common genes

```{r,eval=TRUE}
common_genes <-  intersect(rownames(sce_ms),rownames(sce_ad))
```

```{r,eval=TRUE}
length(common_genes)
```

```{r,eval=TRUE}
sce_ms <- sce_ms[common_genes,]
sce_ad <- sce_ad[common_genes,]
```

Save ensembl - symbol names from ms dataset

```{r,eval=TRUE}
ensembl2symbol <- rowData(sce_ms) %>% 
  as_tibble() %>% 
  dplyr::select(-type) %>% 
  write_tsv('data_sensitive/expression/ensembl2symbol_ms.txt')
```

# Get covariates

For all samples, get:

1. Original dataset (Roche_AD,Columbia_AD,Mathys_AD,Zhou,MS)
2. Case-control status

```{r,eval=TRUE}
ad_cov <- colData(sce_ad) %>% as_tibble() %>% 
  dplyr::select(individual_id,study,diagnosis) %>% 
  unique() %>% 
  mutate(study=paste0(study,'_AD'))
```

```{r,eval=TRUE}
ms_cov <- colData(sce_ms) %>% as_tibble() %>% 
dplyr::select(individual_id,diagnosis=disease_status) %>% 
  mutate(study='MS') %>% 
  unique() %>% 
  mutate(diagnosis=case_when(
    diagnosis=='CTR' ~ 'Ctrl',
    TRUE ~ 'MS'
  )) %>% unique()
```

```{r,eval=TRUE}
cov_all <- rbind(ad_cov,ms_cov)
```

```{r,eval=TRUE}
cov_all_t <- cov_all %>% column_to_rownames('individual_id') %>% t() %>% as.data.frame() %>% rownames_to_column('id')
```

Load genotype PCs

```{r,eval=TRUE,message=FALSE}
cov_genotype_pcs <- read_tsv('data_sensitive/genotypes/pca_covariate_fastqtl.txt')
```

```{r,eval=TRUE}
cov_all_t <- cov_all_t[,colnames(cov_genotype_pcs)]
```

Add covariates from metadata (study, case-control) to genotype PCs.

```{r,eval=TRUE}
cov <- rbind(cov_genotype_pcs,cov_all_t) %>% as_tibble()
```

```{r,eval=TRUE}
dir.create('data_sensitive/eqtl',showWarnings = FALSE)
write_tsv(cov,'data_sensitive/eqtl/covariate_pca_meta_fastqtl.txt')
```

# Get Pseudo-bulk expression function

```{r,eval=TRUE}
count_per_cluster <- function(sce,i,var){
  cnt_matrix <- sce[,sce$broad_annot==cell_type_id$V1[i] & sce[[var]]==cell_type_id$V2[i]]
  tot_counts <- Matrix::rowSums(counts(cnt_matrix))
  n_expressed <- Matrix::rowSums(counts(cnt_matrix)>0)
  n_cells <- ncol(cnt_matrix)
  cnt_df <- data.frame(counts=tot_counts,
                       n_expressed=n_expressed,
                       n_cells=n_cells) %>% 
    rownames_to_column('ensembl') %>% 
    mutate(perc_expressed=round(n_expressed*100/n_cells,digits=2),
           cell_type=cell_type_id$V1[i],
           individual_id=cell_type_id$V2[i]) %>% 
    inner_join(.,ensembl2symbol,by='ensembl') %>% 
  dplyr::select(cell_type,individual_id,ensembl,symbol,counts,n_expressed,n_cells,perc_expressed)
  
  if(var=='sample_id'){
    colnames(cnt_df)[2] <- 'sample_id'
  }
  return(cnt_df)
}
```

## MS

```{r,eval=TRUE}
if(!file.exists('data_sensitive/expression/ms_sum_expression.individual_id.rds')){
  
  cell_type_id <- cbind(as.character(sce_ms$broad_annot),sce_ms$individual_id) %>% unique() %>% as.data.frame()
  
  sum_expression_ms <- mclapply(1:nrow(cell_type_id),
                                count_per_cluster,
                                sce=sce_ms,
                                var='individual_id',
                                mc.cores = 36,
                                mc.preschedule = FALSE) %>% 
    bind_rows()
  
  saveRDS(sum_expression_ms,'data_sensitive/expression/ms_sum_expression.individual_id.rds')
  rm(sum_expression_ms)
  gc()
}
```

```{r,eval=TRUE}
if(!file.exists('data_sensitive/expression/ms_sum_expression.sample_id.rds')){

  sce_ms$sample_id <- sce_ms$library_id
  cell_type_id <- cbind(as.character(sce_ms$broad_annot),sce_ms$sample_id) %>% unique() %>% as.data.frame()
  
  sum_expression_ms <- mclapply(1:nrow(cell_type_id),
                                count_per_cluster,
                                sce=sce_ms,
                                var='sample_id',
                                mc.cores = 36,
                                mc.preschedule = FALSE) %>% 
    bind_rows()
  
  saveRDS(sum_expression_ms,'data_sensitive/expression/ms_sum_expression.sample_id.rds')
  rm(sum_expression_ms)
  gc()
}
```

```{r,eval=TRUE}
rm(sce_ms)
gc()
```

## AD

```{r,eval=TRUE}
if(!file.exists('data_sensitive/expression/ad_sum_expression.individual_id.rds')){
  
  cell_type_id <- cbind(as.character(sce_ad$broad_annot),sce_ad$individual_id) %>% unique() %>% as.data.frame()
  
  sum_expression_ad <- mclapply(1:nrow(cell_type_id),
                                count_per_cluster,
                                sce=sce_ad,
                                var='individual_id',
                                mc.cores = 36,
                                mc.preschedule = FALSE) %>% 
    bind_rows()
  
  saveRDS(sum_expression_ad,'data_sensitive/expression/ad_sum_expression.individual_id.rds')
  rm(sum_expression_ad)
  gc()
}
```

```{r,eval=TRUE}
if(!file.exists('data_sensitive/expression/ad_sum_expression.sample_id.rds')){
  
  cell_type_id <- cbind(as.character(sce_ad$broad_annot),sce_ad$sample_id) %>% unique() %>% as.data.frame()
  
  sum_expression_ad <- mclapply(1:nrow(cell_type_id),
                                count_per_cluster,
                                sce=sce_ad,
                                var='sample_id',
                                mc.cores = 36,
                                mc.preschedule = FALSE) %>% 
    bind_rows()
  
  saveRDS(sum_expression_ad,'data_sensitive/expression/ad_sum_expression.sample_id.rds')
  rm(sum_expression_ad)
  gc()
}
```

```{r,eval=TRUE}
rm(sce_ad)
gc()
```

# Load Pseudo-bulk

```{r,eval=TRUE}
sum_expression_ms <- readRDS('data_sensitive/expression/ms_sum_expression.individual_id.rds') %>% mutate(dataset='ms')
sum_expression_ad <- readRDS('data_sensitive/expression/ad_sum_expression.individual_id.rds') %>% mutate(dataset='ad')
```

## Aggregate datasets

Let's merge the MS and AD dataset.

```{r,eval=TRUE}
sum_expression <- rbind(sum_expression_ms,sum_expression_ad)
```

Get counts per million (CPM)

```{r,eval=TRUE}
sum_expression <- sum_expression %>% 
  group_by(cell_type,individual_id) %>% 
  mutate(counts_scaled_1M=counts*10^6/sum(counts)) %>% 
  ungroup() %>% 
  as_tibble()
```

Save expression of cell type marker genes

```{r,eval=TRUE}
markers <- sum_expression %>% filter(symbol%in%c('SLC17A7','GAD2','AQP4','MOG','PDGFRA','C1QA','CLDN5','RGS5'))
saveRDS(markers,'data_sensitive/expression/cell_marker_expression.rds')
```

## Get number of samples expressing genes and average expression

```{r,eval=TRUE,message=FALSE}
sumstats_gene <- sum_expression %>% group_by(cell_type,ensembl) %>% 
  summarise(counts_per_celltype=sum(counts),
         n_expressed_per_celltype=sum(n_expressed),
         n_cells_per_celltype=sum(n_cells),
         perc_expressed_per_celltype=round(n_expressed_per_celltype*100/n_cells_per_celltype,digits=2),
         n_samples_expressed=sum(counts>0),
         perc_samples_expressed=round(n_samples_expressed*100/length(counts),digits=2),
         mean_cpm=mean(counts_scaled_1M)) %>% 
  group_by(ensembl) %>% 
  mutate(prop_max_cpm=mean_cpm/max(mean_cpm)) %>% 
  ungroup()
```

## Filter genes

### Get Soupy genes

We will first compute the proportion of reads that could be attributed to ambient RNA (conservative estimate).

```{r,eval=TRUE}
compute_ambient_per_sample <- function(id){
  message(id)
  #Get only counts for sample 'id'
  counts_id <- counts_per_cell_type %>% 
    filter(sample_id==id)
  
  #Get the cell types for which we have counts in sample 'id'
  cell_types_id <- counts_id %>% pull(cell_type) %>% unique()
  
  #Get only ambient RNA for sample 'id'
  ambient_id <- ambient %>% filter(sample_id==id) %>% 
    arrange(ensembl)
  
  #Compute ambient proportion for each gene in each cell type from sample 'id'
  ambient_prop <- lapply(1:length(cell_types_id), function(i){
    message(cell_types_id[i])
    counts_id_cell_type <- filter(counts_id,cell_type==cell_types_id[i]) %>% 
    arrange(ensembl)
    stopifnot(all(counts_id_cell_type$ensembl==ambient_id$ensembl))

    max_ambience <- DropletUtils::maximumAmbience(y = counts_id_cell_type$counts,
                                                  ambient = ambient_id$counts,
                                                  mode='proportion')
    out <- tibble(individual_id=id,cell_type=cell_types_id[i],ensembl=counts_id_cell_type$ensembl,soup=max_ambience)
  }) %>% 
    bind_rows()
  
  return(ambient_prop)
}
```

```{r,eval=TRUE}
if(file.exists('data_sensitive/ambient/ambient.prop.estimate.droplet_utils.rds')){
  ambient_prop_per_cell_type <- readRDS('data_sensitive/ambient/ambient.prop.estimate.droplet_utils.rds')
} else{

  #Load pseudo-bulk counts for each sample in the dataset.
  sample_id_ms <- readRDS('data_sensitive/expression/ms_sum_expression.sample_id.rds')
  sample_id_ad <- readRDS('data_sensitive/expression/ad_sum_expression.sample_id.rds')
  
  counts_per_cell_type <- rbind(sample_id_ms,sample_id_ad) %>% 
    dplyr::select(cell_type,sample_id,ensembl,counts)
  
  #We also summed counts for all barcodes with less than 100UMI
  ambient_ad <- read_tsv('data_sensitive/ambient/ambient.100UMI.ad.txt')

  #Swap two samples (genotype of scRNA-seq matches another sample and vice-versa), already done in ad sce object
  colnames(ambient_ad)[colnames(ambient_ad)=='DWM-B3-24-Cog4-Path1-M'] <- 'tmp'
  colnames(ambient_ad)[colnames(ambient_ad)=='DWM-B3-22-Cog4-Path0-M'] <- 'DWM-B3-24-Cog4-Path1-M'
  colnames(ambient_ad)[colnames(ambient_ad)=='tmp'] <- 'DWM-B3-22-Cog4-Path0-M'
  
  #Correct scRNA-seq sample names (scRNA-seq was not matching metadata)
  colnames(ambient_ad)[colnames(ambient_ad)=='MFC-B2-10-Cog1-Path1-F'] <- 'MFC-B2-10-Cog1-Path0-M'
  colnames(ambient_ad)[colnames(ambient_ad)=='MFC-B2-11-Cog1-Path1-M'] <- 'MFC-B2-11-Cog1-Path1-F'

  #Make the ad ambient data long
  ambient_ad <- ambient_ad %>% 
    gather(sample_id,counts,-gene) %>% 
    mutate(ensembl=gsub('\\..+','',gene)) %>% 
    dplyr::select(-gene)
  
  #Now get ambient RNA for ms samples
  ambient_ms <- read_tsv('data_sensitive/ambient/ambient.100UMI.ms.txt')

  #Swap two samples (genotype of scRNA-seq matches another sample and vice-versa)
  colnames(ambient_ms)[colnames(ambient_ms)=='EU017'] <- 'tmp'
  colnames(ambient_ms)[colnames(ambient_ms)=='EU018'] <- 'EU017'
  colnames(ambient_ms)[colnames(ambient_ms)=='tmp'] <- 'EU018'
  
  colnames(ambient_ms)[colnames(ambient_ms)=='WM189'] <- 'tmp'
  colnames(ambient_ms)[colnames(ambient_ms)=='WM190'] <- 'WM189'
  colnames(ambient_ms)[colnames(ambient_ms)=='tmp'] <- 'WM190'
  
  #Make the ms ambient data long
  ambient_ms <- ambient_ms %>% 
    dplyr::select(-symbol) %>% 
    gather(sample_id,counts,-ensembl)
  
  #Aggregate ambient counts for ad and MS, only keep genes that are in the pseudo-bulks counts
  ambient <- rbind(ambient_ad,ambient_ms) %>% 
    filter(ensembl%in%unique(counts_per_cell_type$ensembl))
  
  samples <- unique(counts_per_cell_type$sample_id)
  samples_ambient <- unique(ambient$sample_id)
  stopifnot(all(samples%in%samples_ambient))
  
  #Estimate ambient RNA for each individual
  ambient_prop_all <- mclapply(samples,compute_ambient_per_sample,mc.cores = 16) %>% 
  bind_rows()
  saveRDS(ambient_prop_all,'data_sensitive/ambient/ambient.prop.estimate.droplet_utils.all.rds')

  
  ambient_prop_per_cell_type <- ambient_prop_all%>% 
  group_by(cell_type,ensembl) %>% 
  summarise(mean_soup=mean(soup,na.rm=TRUE),
            median_soup=median(soup,na.rm=TRUE),
            sd_soup=sd(soup,na.rm=TRUE),
            min_soup=min(soup,na.rm=TRUE),
            max_soup=max(soup,na.rm=TRUE),
            n_sample=n())
  
  saveRDS(ambient_prop_per_cell_type,'data_sensitive/ambient/ambient.prop.estimate.droplet_utils.rds')
}
```

### Filtering

Keep genes:

1) For which ambient RNA does not explain more than 10% of the read counts.
2) Expressed in at least 10 individuals 
3) Expressed with a mean CPM of at least 1

```{r,eval=TRUE}
ambient_keep <- filter(ambient_prop_per_cell_type,mean_soup<0.1)
```

```{r,eval=TRUE}
gene_to_keep_per_cell_type <- sumstats_gene %>% 
  filter(n_samples_expressed>10,mean_cpm>1) %>% 
  dplyr::select(cell_type,ensembl) %>% 
  unique() %>% 
  inner_join(.,ambient_keep,by=c('cell_type','ensembl'))
```

```{r,eval=TRUE}
sum_expression <- sum_expression %>% 
  inner_join(.,gene_to_keep_per_cell_type,by=c('cell_type','ensembl'))
```

Number of genes tested per cell type

```{r,eval=TRUE}
(dplyr::count(sum_expression,cell_type,individual_id) %>% 
  dplyr::select(-individual_id) %>% 
  unique() %>% 
  arrange(-n))
```

## Filter individuals

Remove individuals with less than 10 cells for a given cell type.

```{r,eval=TRUE}
(n_cells_ind_cell_type <- sum_expression %>% 
  dplyr::select(cell_type,individual_id,n_cells) %>% 
  unique() %>% arrange(n_cells))
```

```{r,eval=TRUE}
keep <- filter(n_cells_ind_cell_type,n_cells>=10)
```

Number of unique individuals per cell types for eQTL analysis

```{r,eval=TRUE}
dplyr::count(keep,cell_type) %>% arrange(-n)
```

```{r,eval=TRUE}
sum_expression <- inner_join(sum_expression,keep,by=c('cell_type','individual_id','n_cells'))
```

## Get gene coordinates

Get TSS coordinate for all genes.

```{r,eval=TRUE}
gtf <- rtracklayer::import('data/gencode/Homo_sapiens.GRCh38.96.filtered.gtf') %>% 
  as.data.frame() %>% 
  dplyr::filter(type=='gene') %>% 
  mutate(TSS_start=ifelse(strand=='+',start,end),
         TSS_end=ifelse(strand=='+',start+1,end+1)) %>% 
  mutate(gene=paste0(gene_name,'_',gene_id)) %>% 
  dplyr::select(seqnames,TSS_start,TSS_end,gene) %>% 
  filter(seqnames%in%c(1:22))
gtf$seqnames <- droplevels(gtf$seqnames)
```

Set all samples to the same set of symbol and gene ID (like MS dataset).

```{r,eval=TRUE}
gene_id_common <- sum_expression %>% 
  filter(dataset=='ms') %>% 
  dplyr::select(symbol,ensembl) %>% 
  unique() %>% 
  mutate(gene=paste0(symbol,'_',ensembl)) %>% dplyr::select(-symbol)
```

```{r,eval=TRUE}
sum_expression <- sum_expression %>% 
  inner_join(.,gene_id_common,by='ensembl') %>% 
  dplyr::select(cell_type,individual_id,gene,counts,n_expressed,n_cells,perc_expressed,counts_scaled_1M)
```

```{r,eval=TRUE}
saveRDS(sum_expression,'data_sensitive/expression/all_sum_expression.rds')
```

## Get fastQTL format

Perform TMM normalization and get fastQTL format

```{r,eval=TRUE}
get_fastqtl_pheno <- function(cell){
  #keep all expression for the cell type and patient that have genotype
  bed <- sum_expression %>% dplyr::filter(cell_type==cell) %>% 
    dplyr::select(individual_id,gene,counts) %>% 
    spread(individual_id,counts) %>% 
    column_to_rownames('gene') %>% 
    edgeR::DGEList(counts = .) %>% #TMM normalize the data using edgeR
    edgeR::calcNormFactors(.) %>% 
    edgeR::cpm(.) %>%
    as.data.frame() %>% 
    rownames_to_column('gene')
  
    bed <- bed %>% inner_join(.,gtf,by='gene') %>% 
      dplyr::select(seqnames,TSS_start,TSS_end,gene,everything()) %>% 
      arrange(seqnames,TSS_start)
    
    colnames(bed)[c(1,2,3,4)] <- c('#Chr','start','end','ID')
    
    write_tsv(bed,path = paste0('data_sensitive/eqtl/',make.names(cell),'.bed'))
}    
```

```{r,eval=TRUE}
cells <- unique(sum_expression$cell_type)
```

```{r,results=FALSE,eval=TRUE}
mclapply(cells,get_fastqtl_pheno,mc.cores=8)
```

```{bash,eval=TRUE}
cd data_sensitive/eqtl
ml tabix/0.2.6-goolf-1.7.20

for f in *.bed
do
bgzip $f && tabix -p bed $f.gz
done
```

## Generate expression pca

```{r,eval=TRUE,message=FALSE}
geno_cov <- read_tsv('data_sensitive/eqtl/covariate_pca_meta_fastqtl.txt')
``` 

```{r,eval=TRUE}
pheno_files <- list.files('data_sensitive/eqtl/',pattern='*.bed.gz$',full.names = T)
```

```{r,eval=TRUE}
get_covariates <- function(i,folder_name){
  
  out_name <- basename(pheno_files[i]) %>% gsub('\\.bed\\.gz','',.) %>% paste0('.cov.txt')
  
  #load expression data
  d <- data.table::fread(pheno_files[i],data.table=FALSE)
  #perform pca
  pca <- d[-c(1,2,3,4)] %>% t() %>% prcomp(.,scale.=T)
  
  #Create directories with different number of expression pcs
  #Add the n first pcs to the covariate file and write the covariate file in fastqtl format
  number_of_pcs <- c(10,20,30,40,50,60,70,80,90,100,110,120)
  
  lapply(number_of_pcs,function(i){
    
      top_pcs <- pca$x[,1:i] %>% t()
      rownames(top_pcs) <- paste0(rownames(top_pcs),'_exp')
      top_pcs <- top_pcs %>% as.data.frame() %>% rownames_to_column('id')
  
      shared_samples <- intersect(colnames(geno_cov),colnames(top_pcs))
  
      cov <- rbind(geno_cov[,shared_samples],top_pcs[,shared_samples])
      dir.create(paste0(folder_name,i),showWarnings = FALSE)
      write_tsv(cov,paste0(folder_name,i,'/',out_name))
  })
}
```

```{r,results=FALSE,eval=TRUE}
mclapply(1:length(pheno_files),get_covariates,folder_name='data_sensitive/eqtl/PC',mc.cores=8)
```

```{bash,eval=TRUE}
cd data_sensitive/eqtl
for f in PC*
do
cd $f
gzip *
ln -s ../*.bed.gz .
ln -s ../*.bed.gz.tbi .
ln -s ../../genotypes/processed/combined_final.vcf.gz .
ln -s ../../genotypes/processed/combined_final.vcf.gz.tbi .
cd ..
done
```

## Pseudo-bulk eQTL

Let's generate pseudo-bulk expression data across all cell types.

First, we load the expression data at pseudo-bulk level for each cell type.

```{r,eval=TRUE,message=FALSE}
sum_expression_ms <- readRDS('data_sensitive/expression/ms_sum_expression.individual_id.rds') %>% mutate(dataset='ms')
sum_expression_ad <- readRDS('data_sensitive/expression/ad_sum_expression.individual_id.rds') %>% mutate(dataset='ad')
```

```{r,eval=TRUE}
sum_expression <- rbind(sum_expression_ms,sum_expression_ad)
```

Set all samples to the same set of symbol and gene ID (like MS dataset).

```{r,eval=TRUE}
gene_id_common <- sum_expression %>% 
  filter(dataset=='ms') %>% 
  dplyr::select(symbol,ensembl) %>% 
  unique() %>% 
  mutate(gene=paste0(symbol,'_',ensembl)) %>% dplyr::select(-symbol)
```

```{r,eval=TRUE}
sum_expression <- sum_expression %>% 
  inner_join(.,gene_id_common,by='ensembl') %>% 
  dplyr::select(cell_type,individual_id,gene,counts)
```

Get pseudo-bulk expression for each individual accross all cell types

```{r,eval=TRUE}
pb <- sum_expression %>% 
  group_by(individual_id,gene) %>% 
  summarise(counts=sum(counts)) %>% 
  ungroup()
```

Only keep genes with a mean CPM > 1

```{r}
pb <- pb %>% group_by(individual_id) %>% 
  mutate(cpm=counts*10^6/sum(counts)) %>% 
  group_by(gene) %>% 
  mutate(mean_cpm=mean(cpm)) %>% 
  ungroup() %>% 
  filter(mean_cpm>1)
```

Perform TMM normalization and get fastQTL format

```{r,eval=TRUE}
dir.create('data_sensitive/eqtl_pb/',showWarnings = FALSE)

bed <- pb %>%
    dplyr::select(individual_id,gene,counts) %>% 
    spread(individual_id,counts) %>% 
    column_to_rownames('gene') %>% 
    edgeR::DGEList(counts = .) %>% #TMM normalize the data using edgeR
    edgeR::calcNormFactors(.) %>% 
    edgeR::cpm(.) %>%
    as.data.frame() %>% 
    rownames_to_column('gene')

bed <- bed %>% inner_join(.,gtf,by='gene') %>% 
      dplyr::select(seqnames,TSS_start,TSS_end,gene,everything()) %>% 
      arrange(seqnames,TSS_start)
    
colnames(bed)[c(1,2,3,4)] <- c('#Chr','start','end','ID')
    
write_tsv(bed,path = 'data_sensitive/eqtl_pb/pb.bed')
```

```{bash,eval=TRUE}
cd data_sensitive/eqtl_pb
ml tabix/0.2.6-goolf-1.7.20

for f in *.bed
do
bgzip $f && tabix -p bed $f.gz
done
```

## Generate expression pca

```{r,eval=TRUE}
pheno_files <- list.files('data_sensitive/eqtl_pb/',pattern='*.bed.gz$',full.names = T)
```

```{r,eval=TRUE}
get_covariates <- function(i,folder_name){
  
  out_name <- basename(pheno_files[i]) %>% gsub('\\.bed\\.gz','',.) %>% paste0('.cov.txt')
  
  #load expression data
  d <- data.table::fread(pheno_files[i],data.table=FALSE)
  #perform pca
  pca <- d[-c(1,2,3,4)] %>% t() %>% prcomp(.,scale.=T)
  
  #Create directories with different number of expression pcs
  #Add the n first pcs to the covariate file and write the covariate file in fastqtl format
  number_of_pcs <- c(10,20,30,40,50,60,70,80,90,100,110,120)
  
  lapply(number_of_pcs,function(i){
    
      top_pcs <- pca$x[,1:i] %>% t()
      rownames(top_pcs) <- paste0(rownames(top_pcs),'_exp')
      top_pcs <- top_pcs %>% as.data.frame() %>% rownames_to_column('id')
  
      shared_samples <- intersect(colnames(geno_cov),colnames(top_pcs))
  
      cov <- rbind(geno_cov[,shared_samples],top_pcs[,shared_samples])
      dir.create(paste0(folder_name,i),showWarnings = FALSE)
      write_tsv(cov,paste0(folder_name,i,'/',out_name))
  })
}
```

```{r,results=FALSE,eval=TRUE}
lapply(1:length(pheno_files),get_covariates,folder_name='data_sensitive/eqtl_pb/PC')
```

```{bash,eval=TRUE}
cd data_sensitive/eqtl_pb
for f in PC*
do
cd $f
gzip *
ln -s ../*.bed.gz .
ln -s ../*.bed.gz.tbi .
ln -s ../../genotypes/processed/combined_final.vcf.gz .
ln -s ../../genotypes/processed/combined_final.vcf.gz.tbi .
cd ..
done
```
